Move some tests from core to expr
